mac_coordinator/requirements.txt
aiohttp==3.9.1
python-binance==1.0.19
ccxt==4.1.92
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
asyncio==3.4.3
aiosqlite==0.19.0
python-dotenv==1.0.0
ta-lib==0.4.28
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
